requirements.txt
streamlit
pandas
python-dotenv
requests
perplexity-client